@@ -10,7 +10,7 @@
     "assets-webpack-plugin": "^3.5.1",
     "autoprefixer": "^7.2.3",
     "babel-core": "^6.26.3",
-    "babel-eslint": "^6.1.2",
+    "babel-eslint": "^8.2.3",
     "babel-loader": "^7.1.2",
     "babel-preset-env": "^1.7.0",
     "babel-preset-flow": "^6.23.0",
@@ -31,13 +31,13 @@
     "enzyme-adapter-react-16": "^1.1.1",
     "es6-promise": "^4.0.5",
     "es6-promisify": "^5.0.0",
-    "eslint": "^3.11.1",
-    "eslint-config-airbnb": "^12.0.0",
-    "eslint-config-prettier": "^2.1.1",
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-config-prettier": "^2.9.0",
     "eslint-plugin-flowtype": "^2.29.1",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^2.2.2",
-    "eslint-plugin-react": "^6.2.1",
+    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.9.1",
     "extract-text-webpack-plugin": "3.0.2",
     "file-loader": "^1.1.6",
     "flow-bin": "^0.68.0",
@@ -93,7 +93,7 @@
     "sinon": "^4.5.0",
     "style-loader": "^0.19.1",
     "svg-inline-loader": "^0.8.0",
-    "svg-inline-react": "^3.0.0",
+    "svg-inline-react": "^3.1.0",
     "svgo": "^0.7.2",
     "twitter": "^1.7.0",
     "underscore-template-loader": "^0.8.0",
@@ -122,5 +122,8 @@
         "make test"
       ]
     }
+  },
+  "release": {
+    "analyzeCommits": "simple-commit-message"
   }
 }
