@@ -1,6 +1,6 @@
 import assetsDigest from './client-digest';
 
-const bundleName = assetsDigest.metadata.bundleName;
+const { bundleName } = assetsDigest.metadata;
 const assets = assetsDigest[bundleName];
 
 export const cssPath = assets.css;
