@@ -4,26 +4,21 @@ import * as React from 'react';
 import styles from './style.css';
 
 type CellProps = {
-  breakOn: string,
+  breakOn?: string,
   children?: React.Node,
-  hideOn: 'mobileS' | 'mobileSM' | 'mobile' | 'tablet' | 'dont',
+  hideOn?: 'mobileS' | 'mobileSM' | 'mobile' | 'tablet' | 'dont',
   size: number,
 };
 
-export default class Cell extends React.Component<CellProps> {
-  static defaultProps = {
-    breakOn: 'mobile',
-    hideOn: 'dont',
-  };
+const Cell = ({ breakOn = 'mobile', children, hideOn = 'dont', size }: CellProps) => {
+  const cellClassNames = classNames({
+    [styles[`responsive-cell-${breakOn}`]]: true,
+    [styles.cell]: true,
+    [styles[`size${size}of12`]]: !!size,
+    [styles[`hideOn${hideOn}`]]: true,
+  });
 
-  render() {
-    const cellClassNames = classNames({
-      [styles[`responsive-cell-${this.props.breakOn}`]]: true,
-      [styles.cell]: true,
-      [styles[`size${this.props.size}of12`]]: !!this.props.size,
-      [styles[`hideOn${this.props.hideOn}`]]: true,
-    });
+  return <div className={cellClassNames}>{children}</div>;
+};
 
-    return <div className={cellClassNames}>{this.props.children}</div>;
-  }
-}
+export default Cell;
