@@ -33,8 +33,13 @@ class ClientTestimonialsSlice extends React.Component<*, ClientTestimonialState>
             <Navigator currentIndex={currentIndex} onClick={this.setPage} />
             <div className={styles.content}>
               <SwipeableViews index={currentIndex} onChangeIndex={this.setPage}>
-                {testimonials.map((t, i) => (
-                  <Testimonial key={i} type={t.type} content={t.content} author={t.author} />
+                {testimonials.map(t => (
+                  <Testimonial
+                    key={t.content}
+                    type={t.type}
+                    content={t.content}
+                    author={t.author}
+                  />
                 ))}
               </SwipeableViews>
             </div>
