@@ -43,7 +43,7 @@ const FMTeaCaseStudy = () => (
           </h1>
         </div>
         <h2 className={styles.content__title}>
-          <span className={styles.content__redTitle}>{"Let's make things better."}</span>
+          <span className={styles.content__redTitle}>Let&apos;s make things better.</span>
           Make everyday special
         </h2>
         <p className={styles.content__paragraph}>
@@ -108,7 +108,7 @@ const FMTeaCaseStudy = () => (
           instantly.
         </p>
         <h2 className={styles.content__title}>
-          <span className={styles.content__redTitle}>{'Creating lasting change.'}</span>
+          <span className={styles.content__redTitle}>Creating lasting change.</span>
           Learning fast and scaling excellence
         </h2>
         <p className={styles.content__paragraph}>
