@@ -1,17 +1,21 @@
 export default {
   roundtable: {
-    __html: `<!--HubSpot Call-to-Action Code --><span class="hs-cta-wrapper" id="hs-cta-wrapper-77faffb7-d221-4ae0-a31b-bfa4ea9816c0"><span class="hs-cta-node hs-cta-77faffb7-d221-4ae0-a31b-bfa4ea9816c0" id="hs-cta-77faffb7-d221-4ae0-a31b-bfa4ea9816c0"><!--[if lte IE 8]><div id="hs-cta-ie-element"></div><![endif]--><a href="https://cta-redirect.hubspot.com/cta/redirect/4210858/77faffb7-d221-4ae0-a31b-bfa4ea9816c0"  target="_blank" ><img class="hs-cta-img" id="hs-cta-img-77faffb7-d221-4ae0-a31b-bfa4ea9816c0" style="border-width:0px;" height="112" width="456" src="https://no-cache.hubspot.com/cta/default/4210858/77faffb7-d221-4ae0-a31b-bfa4ea9816c0.png"  alt="Read PDF report"/></a></span><script charset="utf-8" src="https://js.hscta.net/cta/current.js"></script><script type="text/javascript"> hbspt.cta.load(4210858, '77faffb7-d221-4ae0-a31b-bfa4ea9816c0', {}); </script></span><!-- end HubSpot Call-to-Action Code -->`,
+    __html:
+      '<!--HubSpot Call-to-Action Code --><span class="hs-cta-wrapper" id="hs-cta-wrapper-77faffb7-d221-4ae0-a31b-bfa4ea9816c0"><span class="hs-cta-node hs-cta-77faffb7-d221-4ae0-a31b-bfa4ea9816c0" id="hs-cta-77faffb7-d221-4ae0-a31b-bfa4ea9816c0"><!--[if lte IE 8]><div id="hs-cta-ie-element"></div><![endif]--><a href="https://cta-redirect.hubspot.com/cta/redirect/4210858/77faffb7-d221-4ae0-a31b-bfa4ea9816c0"  target="_blank" ><img class="hs-cta-img" id="hs-cta-img-77faffb7-d221-4ae0-a31b-bfa4ea9816c0" style="border-width:0px;" height="112" width="456" src="https://no-cache.hubspot.com/cta/default/4210858/77faffb7-d221-4ae0-a31b-bfa4ea9816c0.png"  alt="Read PDF report"/></a></span><script charset="utf-8" src="https://js.hscta.net/cta/current.js"></script><script type="text/javascript"> hbspt.cta.load(4210858, \'77faffb7-d221-4ae0-a31b-bfa4ea9816c0\', {}); </script></span><!-- end HubSpot Call-to-Action Code -->',
   },
 
   slack: {
-    __html: `<!--HubSpot Call-to-Action Code --><span class="hs-cta-wrapper" id="hs-cta-wrapper-b81465b3-9974-475f-992a-d90d6a90b275"><span class="hs-cta-node hs-cta-b81465b3-9974-475f-992a-d90d6a90b275" id="hs-cta-b81465b3-9974-475f-992a-d90d6a90b275"><!--[if lte IE 8]><div id="hs-cta-ie-element"></div><![endif]--><a href="https://cta-redirect.hubspot.com/cta/redirect/4210858/b81465b3-9974-475f-992a-d90d6a90b275"  target="_blank" ><img class="hs-cta-img" id="hs-cta-img-b81465b3-9974-475f-992a-d90d6a90b275" style="border-width:0px;" height="112" width="502" src="https://no-cache.hubspot.com/cta/default/4210858/b81465b3-9974-475f-992a-d90d6a90b275.png"  alt="Join us on slack"/></a></span><script charset="utf-8" src="https://js.hscta.net/cta/current.js"></script><script type="text/javascript"> hbspt.cta.load(4210858, 'b81465b3-9974-475f-992a-d90d6a90b275', {}); </script></span><!-- end HubSpot Call-to-Action Code -->`,
+    __html:
+      '<!--HubSpot Call-to-Action Code --><span class="hs-cta-wrapper" id="hs-cta-wrapper-b81465b3-9974-475f-992a-d90d6a90b275"><span class="hs-cta-node hs-cta-b81465b3-9974-475f-992a-d90d6a90b275" id="hs-cta-b81465b3-9974-475f-992a-d90d6a90b275"><!--[if lte IE 8]><div id="hs-cta-ie-element"></div><![endif]--><a href="https://cta-redirect.hubspot.com/cta/redirect/4210858/b81465b3-9974-475f-992a-d90d6a90b275"  target="_blank" ><img class="hs-cta-img" id="hs-cta-img-b81465b3-9974-475f-992a-d90d6a90b275" style="border-width:0px;" height="112" width="502" src="https://no-cache.hubspot.com/cta/default/4210858/b81465b3-9974-475f-992a-d90d6a90b275.png"  alt="Join us on slack"/></a></span><script charset="utf-8" src="https://js.hscta.net/cta/current.js"></script><script type="text/javascript"> hbspt.cta.load(4210858, \'b81465b3-9974-475f-992a-d90d6a90b275\', {}); </script></span><!-- end HubSpot Call-to-Action Code -->',
   },
 
   meetup: {
-    __html: `<!--HubSpot Call-to-Action Code --><span class="hs-cta-wrapper" id="hs-cta-wrapper-d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e"><span class="hs-cta-node hs-cta-d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e" id="hs-cta-d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e"><!--[if lte IE 8]><div id="hs-cta-ie-element"></div><![endif]--><a href="https://cta-redirect.hubspot.com/cta/redirect/4210858/d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e"  target="_blank" ><img class="hs-cta-img" id="hs-cta-img-d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e" style="border-width:0px;" height="112" width="604" src="https://no-cache.hubspot.com/cta/default/4210858/d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e.png"  alt="Come"/></a></span><script charset="utf-8" src="https://js.hscta.net/cta/current.js"></script><script type="text/javascript"> hbspt.cta.load(4210858, 'd9b3a83a-f5d9-491e-9720-c9ab3e6dd09e', {}); </script></span><!-- end HubSpot Call-to-Action Code -->`,
+    __html:
+      '<!--HubSpot Call-to-Action Code --><span class="hs-cta-wrapper" id="hs-cta-wrapper-d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e"><span class="hs-cta-node hs-cta-d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e" id="hs-cta-d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e"><!--[if lte IE 8]><div id="hs-cta-ie-element"></div><![endif]--><a href="https://cta-redirect.hubspot.com/cta/redirect/4210858/d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e"  target="_blank" ><img class="hs-cta-img" id="hs-cta-img-d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e" style="border-width:0px;" height="112" width="604" src="https://no-cache.hubspot.com/cta/default/4210858/d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e.png"  alt="Come"/></a></span><script charset="utf-8" src="https://js.hscta.net/cta/current.js"></script><script type="text/javascript"> hbspt.cta.load(4210858, \'d9b3a83a-f5d9-491e-9720-c9ab3e6dd09e\', {}); </script></span><!-- end HubSpot Call-to-Action Code -->',
   },
 
   webinar: {
-    __html: `<!--HubSpot Call-to-Action Code --><span class="hs-cta-wrapper" id="hs-cta-wrapper-c87452af-5599-4904-a20c-7d9d48886193"><span class="hs-cta-node hs-cta-c87452af-5599-4904-a20c-7d9d48886193" id="hs-cta-c87452af-5599-4904-a20c-7d9d48886193"><!--[if lte IE 8]><div id="hs-cta-ie-element"></div><![endif]--><a href="https://cta-redirect.hubspot.com/cta/redirect/4210858/c87452af-5599-4904-a20c-7d9d48886193"  target="_blank" ><img class="hs-cta-img" id="hs-cta-img-c87452af-5599-4904-a20c-7d9d48886193" style="border-width:0px;" height="112" width="448" src="https://no-cache.hubspot.com/cta/default/4210858/c87452af-5599-4904-a20c-7d9d48886193.png"  alt="Watch them now"/></a></span><script charset="utf-8" src="https://js.hscta.net/cta/current.js"></script><script type="text/javascript"> hbspt.cta.load(4210858, 'c87452af-5599-4904-a20c-7d9d48886193', {}); </script></span><!-- end HubSpot Call-to-Action Code -->`,
+    __html:
+      '<!--HubSpot Call-to-Action Code --><span class="hs-cta-wrapper" id="hs-cta-wrapper-c87452af-5599-4904-a20c-7d9d48886193"><span class="hs-cta-node hs-cta-c87452af-5599-4904-a20c-7d9d48886193" id="hs-cta-c87452af-5599-4904-a20c-7d9d48886193"><!--[if lte IE 8]><div id="hs-cta-ie-element"></div><![endif]--><a href="https://cta-redirect.hubspot.com/cta/redirect/4210858/c87452af-5599-4904-a20c-7d9d48886193"  target="_blank" ><img class="hs-cta-img" id="hs-cta-img-c87452af-5599-4904-a20c-7d9d48886193" style="border-width:0px;" height="112" width="448" src="https://no-cache.hubspot.com/cta/default/4210858/c87452af-5599-4904-a20c-7d9d48886193.png"  alt="Watch them now"/></a></span><script charset="utf-8" src="https://js.hscta.net/cta/current.js"></script><script type="text/javascript"> hbspt.cta.load(4210858, \'c87452af-5599-4904-a20c-7d9d48886193\', {}); </script></span><!-- end HubSpot Call-to-Action Code -->',
   },
 };
