@@ -14,20 +14,18 @@ const headerSlice = () => {
             <div className={styles.flexibleUpper}>
               <div className={styles.pageHeaderUpper}>
                 {/* Space added so that h1 can be crawled correctly */}
-                Do the{` `}
-                {/* Space added so that h1 can be crawled correctly */}
+                Do the {/* Space added so that h1 can be crawled correctly */}
                 <br />right thing{' '}
               </div>
-              <img alt="" role="presentation" className={styles.arrowDown} src={arrowImg} />
+              <img alt="" className={styles.arrowDown} src={arrowImg} />
             </div>
           </div>
           <div className={styles.alignRight} role="presentation">
             <div className={styles.flexibleLower}>
-              <img alt="" role="presentation" className={styles.arrowUp} src={arrowImg} />
+              <img alt="" className={styles.arrowUp} src={arrowImg} />
               <div className={styles.pageHeaderLower}>
                 {/* Space added so that h1 can be crawled correctly */}
-                Do the{` `}
-                <br />thing right
+                Do the <br />thing right
               </div>
             </div>
           </div>
